chore: updated tnc modal logic
